feat(controllers): Updating sample components demonstrating usage of SlotController.
