+++ src/components/examples/slot-controller/example-slot-controller-base/example-slot-controller-base.css
@@ -1,4 +1,4 @@
-.example-slot-controller-component {
+.example-slot-controller {
   @apply my-8 rounded-xl border-2 border-dashed border-demo-darkBlue bg-demo-lightBlue p-12 text-lg shadow-xl;
 }
